requirements.txt
appdirs==1.4.3
asgiref==3.2.7
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
distlib==0.3.0
Django==3.0.7
djangorestframework==3.11.0
filelock==3.0.12
idna==2.9
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.4
parso==0.7.0
pickleshare==0.7.5
pipenv==2018.11.26
prompt-toolkit==3.0.5
Pygments==2.6.1
pyparsing==2.4.7
pyperclip==1.8.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
termcolor==1.1.0
traitlets==4.3.3
urllib3==1.25.9
virtualenv==20.0.20
virtualenv-clone==0.5.4
wcwidth==0.1.9
